* commit 'ce7a7a3b7b69f1f3638be24b4e24ce1e0bd10b3c':
  NativeCollation LP64 fixes.

+++ luni/src/main/java/libcore/icu/NativeCollation.java
@@ -23,7 +23,7 @@ public final class NativeCollation {
     public static native void closeCollator(long address);
     public static native int compare(long address, String source, String target);
     public static native int getAttribute(long address, int type);
-    public static native int getCollationElementIterator(long address, String source);
+    public static native long getCollationElementIterator(long address, String source);
     public static native String getRules(long address);
     public static native byte[] getSortKey(long address, String source);
     public static native long openCollator(String locale);
